impl parity conversion
