--- crates/revm/revm-inspectors/src/tracing/call.rs
@@ -1,120 +1,0 @@
-use crate::tracing::arena::CallTraceStep;
-use reth_primitives::{bytes::Bytes, rpc::ActionType, Address, U256};
-use reth_rpc_types::trace::parity::CallType;
-use revm::interpreter::{CallContext, CallScheme, CreateScheme, InstructionResult};
-use serde::{Deserialize, Serialize};
-
-/// A trace of a call.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub(crate) struct CallTrace {
-    /// The depth of the call
-    pub(crate) depth: usize,
-    /// Whether the call was successful
-    pub(crate) success: bool,
-    /// caller of this call
-    pub(crate) caller: Address,
-    /// The destination address of the call or the address from the created contract
-    pub(crate) address: Address,
-    /// The kind of call this is
-    pub(crate) kind: CallKind,
-    /// The value transferred in the call
-    pub(crate) value: U256,
-    /// The calldata for the call, or the init code for contract creations
-    pub(crate) data: Bytes,
-    /// The return data of the call if this was not a contract creation, otherwise it is the
-    /// runtime bytecode of the created contract
-    pub(crate) output: Bytes,
-    /// The gas cost of the call
-    pub(crate) gas_cost: u64,
-    /// The status of the trace's call
-    pub(crate) status: InstructionResult,
-    /// call context of the runtime
-    pub(crate) call_context: Option<CallContext>,
-    /// Opcode-level execution steps
-    pub(crate) steps: Vec<CallTraceStep>,
-}
-
-impl CallTrace {
-    /// Whether this is a contract creation
-    pub(crate) fn is_created(&self) -> bool {
-        matches!(self.kind, CallKind::Create | CallKind::Create2)
-    }
-}
-
-impl Default for CallTrace {
-    fn default() -> Self {
-        Self {
-            depth: Default::default(),
-            success: Default::default(),
-            caller: Default::default(),
-            address: Default::default(),
-            kind: Default::default(),
-            value: Default::default(),
-            data: Default::default(),
-            output: Default::default(),
-            gas_cost: Default::default(),
-            status: InstructionResult::Continue,
-            call_context: Default::default(),
-            steps: Default::default(),
-        }
-    }
-}
-
-/// A unified representation of a call
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "UPPERCASE")]
-#[allow(missing_docs)]
-pub enum CallKind {
-    #[default]
-    Call,
-    StaticCall,
-    CallCode,
-    DelegateCall,
-    Create,
-    Create2,
-}
-
-impl From<CallScheme> for CallKind {
-    fn from(scheme: CallScheme) -> Self {
-        match scheme {
-            CallScheme::Call => CallKind::Call,
-            CallScheme::StaticCall => CallKind::StaticCall,
-            CallScheme::CallCode => CallKind::CallCode,
-            CallScheme::DelegateCall => CallKind::DelegateCall,
-        }
-    }
-}
-
-impl From<CreateScheme> for CallKind {
-    fn from(create: CreateScheme) -> Self {
-        match create {
-            CreateScheme::Create => CallKind::Create,
-            CreateScheme::Create2 { .. } => CallKind::Create2,
-        }
-    }
-}
-
-impl From<CallKind> for ActionType {
-    fn from(kind: CallKind) -> Self {
-        match kind {
-            CallKind::Call | CallKind::StaticCall | CallKind::DelegateCall | CallKind::CallCode => {
-                ActionType::Call
-            }
-            CallKind::Create => ActionType::Create,
-            CallKind::Create2 => ActionType::Create,
-        }
-    }
-}
-
-impl From<CallKind> for CallType {
-    fn from(ty: CallKind) -> Self {
-        match ty {
-            CallKind::Call => CallType::Call,
-            CallKind::StaticCall => CallType::StaticCall,
-            CallKind::CallCode => CallType::CallCode,
-            CallKind::DelegateCall => CallType::DelegateCall,
-            CallKind::Create => CallType::None,
-            CallKind::Create2 => CallType::None,
-        }
-    }
-}
